Homework29/src/main/java/com/example/homework29/service/AuthService.java
package com.example.homework29.service;


import com.example.homework29.model.MyUser;
import com.example.homework29.repository.AuthRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class AuthService {

    private final AuthRepository authRepository;


    public void register(MyUser myUser) {
        myUser.setRole("USER");
        String hashedPassword=new BCryptPasswordEncoder().encode(myUser.getPassword());
        myUser.setPassword(hashedPassword);
        authRepository.save(myUser);
    }

    public MyUser getUserById(Integer userId){
        MyUser user= authRepository.findMyUserById(userId);
        return user;

    }

}
